We now have a working, buggy engine to play against!
It can play half decently, but there is definitely a logical error somewhere to do with if it's scoring the current position for itself or its opponent. I'll get my head around it next time.

On an unrelated note, changed tabs -> spaces, meaning I touched almost every line this commit.